# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Armani Duty Free storefront and self-checkout API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: string;
  name: string;
  brand: string;
  category: string;
  price: number;
  compareAtPrice: number;
  image: string;
  description: string;
  size: string;
  badge: string;
  rfidTag: string;
}

export interface CartLineInput {
  productId: string;
  /** @minimum 1 */
  quantity: number;
}

export interface CartQuoteInput {
  items: CartLineInput[];
  destination: string;
  flightTime: string;
}

export interface CartLine {
  product: Product;
  quantity: number;
  lineTotal: number;
}

export interface CartQuote {
  items: CartLine[];
  subtotal: number;
  savings: number;
  duties: number;
  total: number;
  itemCount: number;
  eligible: boolean;
  /**
     * Whole cents available to the signed-in customer
     * @minimum 0
     */
  storeCreditCents: number;
  /**
     * Whole cents applied to this quote
     * @minimum 0
     */
  creditAppliedCents: number;
  message: string;
}

export interface RfidScanInput {
  tagId: string;
  device?: string;
}

export interface RfidScanResult {
  tagId: string;
  found: boolean;
  product: Product | null;
  message: string;
}

export interface CheckoutInput {
  items: CartLineInput[];
  destination: string;
  flightNumber: string;
  email: string;
  paymentMethod: string;
}

export interface CheckoutResult {
  orderId: string;
  total: number;
  collectionPoint: string;
  collectionWindow: string;
  message: string;
}

export interface StoreStatus {
  open: boolean;
  location: string;
  terminal: string;
  collectionPoint: string;
  nextFlight: string;
}

export interface CustomerAccount {
  userId: string;
  /**
     * Whole cents
     * @minimum 0
     */
  storeCreditCents: number;
}

export type GraphqlInputVariables = { [key: string]: unknown };

export interface GraphqlInput {
  query: string;
  variables?: GraphqlInputVariables;
}

export type GraphqlResponseData = { [key: string]: unknown };

export type GraphqlResponseErrorsItem = { [key: string]: unknown };

export interface GraphqlResponse {
  data: GraphqlResponseData;
  errors?: GraphqlResponseErrorsItem[];
}

export type ListCatalogParams = {
category?: string;
search?: string;
};