# lib/api-zod/src/generated/types/cartQuote.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Armani Duty Free storefront and self-checkout API
 * OpenAPI spec version: 0.1.0
 */
import type { CartLine } from './cartLine';

export interface CartQuote {
  items: CartLine[];
  subtotal: number;
  savings: number;
  duties: number;
  total: number;
  itemCount: number;
  eligible: boolean;
  /**
     * Whole cents available to the signed-in customer
     * @minimum 0
     */
  storeCreditCents: number;
  /**
     * Whole cents applied to this quote
     * @minimum 0
     */
  creditAppliedCents: number;
  message: string;
}