# lib/api-zod/src/generated/types/customerAccount.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Armani Duty Free storefront and self-checkout API
 * OpenAPI spec version: 0.1.0
 */

export interface CustomerAccount {
  userId: string;
  /**
     * Whole cents
     * @minimum 0
     */
  storeCreditCents: number;
}